pyproject.toml
[project]
name = "weather-dashboard"
version = "0.1.0"
description = "A dynamic weather dashboard application with Tkinter"
readme = "README.md"
authors = [
    { name = "waleed260", email = "vkdeku20@gmail.com" }
]
requires-python = ">=3.7"
dependencies = [
    "requests>=2.25.0",
    "Pillow>=8.0.0",
    "python-dotenv>=0.19.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.9"
]

[project.scripts]
weather-dashboard = "weather_dashboard.main:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
